08.19 add githubDB

+++ coin/src/Pages/Main/Main.js
@@ -39,7 +39,11 @@ function Main() {
     const _inputData = {
       id: id,
     };
-    dispatch(deleteBoard(_inputData));
+    if (window.confirm("현재 보드를 삭제합니다.")){
+      dispatch(deleteBoard(_inputData));
+      alert("삭제되었습니다.");
+    }
+    
   };
 
   const titleList = board.map((row) => (
